cmd: move discovery service to worker server
