+++ cmd/management_api.go
@@ -1,10 +1,8 @@
 package main
 
 import (
-	"github.com/Shopify/sarama"
 	"github.com/spf13/cobra"
 	"go.od2.network/hive/pkg/auth"
-	"go.od2.network/hive/pkg/discovery"
 	"go.od2.network/hive/pkg/management"
 	"go.od2.network/hive/pkg/types"
 	"go.uber.org/zap"
@@ -55,31 +53,10 @@ func runManagementAPI(cmd *cobra.Command, _ []string) {
 		grpc.UnaryInterceptor(interceptor.Unary()),
 		grpc.StreamInterceptor(interceptor.Stream()),
 	)
-	// Connect to Kafka.
-	saramaClient := saramaClientFromEnv()
-	defer func() {
-		if err := saramaClient.Close(); err != nil {
-			log.Error("Failed to close sarama client", zap.Error(err))
-		}
-	}()
-	log.Info("Creating Kafka producer")
-	producer, err := sarama.NewSyncProducerFromClient(saramaClient)
-	if err != nil {
-		log.Fatal("Failed to build Kafka producer", zap.Error(err))
-	}
-	defer func() {
-		log.Info("Closing Kafka producer")
-		if err := producer.Close(); err != nil {
-			log.Error("Failed to close Kafka producer", zap.Error(err))
-		}
-	}()
 	// Assemble handlers
 	types.RegisterManagementServer(server, &management.Handler{
 		DB:     db.DB,
 		Signer: getSigner(),
-	})
-	types.RegisterDiscoveryServer(server, &discovery.Handler{
-		Producer: producer,
 	})
 	// Start listener
 	listen, err := listenUnix(socket)
